pyproject.toml
[project]
name = "doctor-appointment-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "ipykernel>=6.30.0",
    "ipython>=9.4.0",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
]
